Move ProviderServiceAccount & ServiceDiscovery controller to supervisor package
