+++ controllers/vmware/servicediscovery_controller_unit_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package controllers
+package vmware
 
 import (
 	. "github.com/onsi/ginkgo/v2"
